requirements.txt
tensorflow==2.15.0
torch==2.5.1 
torchvision==0.20.1 
numpy==1.26.4  
opencv-python==4.8.0.76  
ultralytics==8.3.102 
streamlit==1.27.0  
pillow==9.5.0   
pandas==2.2.3 
plotly==5.16.1  
pyyaml==6.0.1  
scikit-learn==1.3.0  
matplotlib==3.7.2    